Change logger inheritance structure to class composition rather than trait extension

+++ gangsutils-logger/src/test/scala/cn/tellyouwhat/gangsutils/logger/dest/fs/LocalPlainTextLoggerTest.scala
@@ -1,7 +1,8 @@
 package cn.tellyouwhat.gangsutils.logger.dest.fs
 
 import cn.tellyouwhat.gangsutils.core.constants.infoHead_unquote
-import cn.tellyouwhat.gangsutils.logger.{GangLogger, SupportedLogDest}
+import cn.tellyouwhat.gangsutils.logger.Logger
+import cn.tellyouwhat.gangsutils.logger.cc.LoggerConfiguration
 import org.scalactic.TimesOnInt.convertIntToRepeater
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
@@ -11,37 +12,35 @@ import java.nio.file.{Files, Paths}
 import scala.io.Source
 
 class LocalPlainTextLoggerTest extends AnyFlatSpec with Matchers with BeforeAndAfter {
+  val path = "log_text/1.txt"
 
   before {
-    GangLogger.resetLoggerConfig()
+    if (Files.exists(Paths.get(path))) {
+      Files.delete(Paths.get(path))
+      Files.delete(Paths.get(path).getParent)
+    }
   }
 
   after {
-    GangLogger.resetLoggerConfig()
-    LocalPlainTextLogger.resetLogSavePath()
+    Files.delete(Paths.get(path))
+    Files.delete(Paths.get(path).getParent)
   }
 
   behavior of "LocalPlainTextLoggerTest"
 
   "setLogSavePath" should "set log save path" in {
-    val path = "/path/to/log.txt"
     LocalPlainTextLogger.setLogSavePath(path)
-    GangLogger().logSavePath shouldBe Paths.get(path)
+    LocalPlainTextLogger.initializeConfiguration(LoggerConfiguration())
+    val logger = LocalPlainTextLogger()
+    logger.asInstanceOf[LocalPlainTextLogger].logSavePath shouldBe Paths.get(path)
   }
 
   "LocalPlainTextLogger" should "log to local file with plain text" in {
-    val path = "loglog/1.txt"
-    LocalPlainTextLogger.setLogSavePath(path)
-    GangLogger.setDefaultLogDest(SupportedLogDest.LOCAL_PLAIN_TEXT_LOGGER :: Nil)
-    GangLogger.disableDateTime()
-    GangLogger.disableHostname()
-    val logger = GangLogger()
+    val logger: Logger = LocalPlainTextLogger(LoggerConfiguration(isDTEnabled = false, isHostnameEnabled = false), path)
     20 times logger.info("hello fs")
     val source = Source.fromFile(path, "UTF-8")
     source.mkString should fullyMatch regex (infoHead_unquote + ": hello fs\\s+") * 20
     source.close()
-    Files.delete(Paths.get(path))
-    Files.delete(Paths.get(path).getParent)
   }
 
 }
